refactor: rename workflow package to dsl
