+++ examples/conditionally-execute/main.go
@@ -22,7 +22,7 @@ import (
 	"math/rand/v2"
 
 	"github.com/mrsimonemms/golang-helpers/temporal"
-	"github.com/mrsimonemms/temporal-dsl/pkg/workflow"
+	"github.com/mrsimonemms/temporal-dsl/pkg/dsl"
 	"github.com/rs/zerolog/log"
 	"go.temporal.io/sdk/client"
 )
@@ -46,7 +46,7 @@ func main() {
 	log.Info().Int("randomInteger", i).Msg("Triggering workflow")
 
 	ctx := context.Background()
-	we, err := c.ExecuteWorkflow(ctx, workflowOptions, "conditional", workflow.HTTPData{
+	we, err := c.ExecuteWorkflow(ctx, workflowOptions, "conditional", dsl.HTTPData{
 		"randomInteger": i,
 	})
 	if err != nil {
@@ -56,7 +56,7 @@ func main() {
 
 	log.Info().Str("workflowId", we.GetID()).Str("runId", we.GetRunID()).Msg("Started workflow")
 
-	var result map[string]workflow.OutputType
+	var result map[string]dsl.OutputType
 	if err := we.Get(ctx, &result); err != nil {
 		log.Fatal().Err(err).Msg("Error getting response")
 	}
