+++ examples/raise/main.go
@@ -20,7 +20,7 @@ import (
 	"context"
 
 	"github.com/mrsimonemms/golang-helpers/temporal"
-	"github.com/mrsimonemms/temporal-dsl/pkg/workflow"
+	"github.com/mrsimonemms/temporal-dsl/pkg/dsl"
 	"github.com/rs/zerolog/log"
 	"go.temporal.io/sdk/client"
 )
@@ -48,7 +48,7 @@ func main() {
 
 	log.Info().Str("workflowId", we.GetID()).Str("runId", we.GetRunID()).Msg("Started workflow")
 
-	var result map[string]workflow.OutputType
+	var result map[string]dsl.OutputType
 	if err := we.Get(ctx, &result); err != nil {
 		log.Fatal().Err(err).Msg("Error getting response")
 	}
